Active tab now dictates the page left menu links to
It got annoying flicking between assets and then having to
select the analytics tab so the idea here is it will
reduce clicks to get to things.

Brakdown page originally was disabled for things that
contained only 1 item .. mostly because it was pointless
however this new functionality bypasses that, so might as well
enable the button for everything!